Add files via upload
This is a Spotify Clone created using HTML, CSS. It replicates the basic design and functionality of Spotify's music player interface. The project includes features like a responsive layout, song list, play/pause controls, and progress bar functionality. It is designed as a frontend-only project to practice DOM manipulation and layout design.